doc: use annotated tag for releases
As documented by `git tag --help`, 'annotated tags are meant for release while lightweight tags are meant for private or temporary object labels'.
Annotated tag are a real commit. It's not the case for lightweight tags.

A shorter command is usable too:
git tag -a XX.YY.ZZ -m "Release version XX.YY.ZZ"

The long parameters are used in the documentation only to be clear about theirs purposes.